Use composite index for Track model
We can't use their id because it can be null.

My implementation for updating tracks triggers EFCore's diff algorithm to
replace the entire collection with fresh ids. This means that there is a
risk of reaching the upper limit on ids when using an integer index. By
employing a composite index of (PlaylistId, PositionInPlaylist) these
problems are resolved.